Use int64_t when 64-bits are essential

+++ config.h
@@ -40,8 +40,8 @@ typedef struct _config {
     char *working_dir;
     uint32_t bind_addr;
     char restart;
-    unsigned int max_text_length;
-    ssize_t max_file_size;
+    uint32_t max_text_length;
+    int64_t max_file_size;
 #ifdef _WIN32
     char tray_icon;
     unsigned short display;
